Customer Churn Data Set for a Telecommunications Company

A dataset containing data related to telecom customers that have enrolled in various products and services

customer_churn_tbl

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 7043 rows and 21 columns.

Source

IBM Sample Datasets

Details

Telecom Customer Data:

customerID (chr): CUSTOMER ID

gender (chr): Customer's gender ("Female", "Male")

SeniorCitizen (dbl): 1 = Senior Citzen, 0 = Not Senior Citizen

Partner (chr): Whether the customer has a partner or not (Yes, No)

Dependents (chr): Whether the customer has dependents or not (Yes, No)

tenure (dbl): Number of months the customer has stayed with the company

PhoneService (chr): Whether the customer has a phone service or not (Yes, No)

MultipleLines (chr): Whether the customer has multiple lines or not (Yes, No, No phone service)

InternetService (chr): Customer’s internet service provider (DSL, Fiber optic, No)

OnlineSecurity (chr): Whether the customer has online security or not (Yes, No, No internet service)

OnlineBackup (chr): Whether the customer has online backup or not (Yes, No, No internet service)

DeviceProtection (chr): Whether the customer has device protection or not (Yes, No, No internet service)

TechSupport (chr): Whether the customer has tech support or not (Yes, No, No internet service)

StreamingTV (chr): Whether the customer has streaming TV or not (Yes, No, No internet service)

StreamingMovies (chr): Whether the customer has streaming movies or not (Yes, No, No internet service)

Contract (chr): The contract term of the customer (Month-to-month, One year, Two year)

PaperlessBilling (chr): Whether the customer has paperless billing or not (Yes, No)

PaymentMethod (chr): The customer’s payment method (Electronic check, Mailed check, Bank transfer (automatic), Credit card (automatic))

MonthlyCharges (dbl): The amount charged to the customer monthly

TotalCharges (dbl): The total amount charged to the customer

Churn (chr): Outcome. Whether the customer churned or not (Yes or No)